Revert "[Misc] main2main 0608(refactor fused moe) (#11081)"
This reverts commit 1c57c6a0a95dd99b06d7d75d3dd014f50f6cdb63.
