+++ .github/vllm-main-verified.commit
@@ -1,1 +1,1 @@
-dc68bd8c4199b00631fe71eb37313f406cc66ac1
+967c5c3bc38891f4465d3f4e99917ed837bb3833
