+++ tests/e2e/pull_request/two_card/test_qwen3_30b_a3b.py
@@ -18,17 +18,10 @@
 
 import json
 
-import pytest
 import requests
 from vllm.utils.network_utils import get_open_port
 
 from tests.e2e.conftest import RemoteOpenAIServer, wait_until_npu_memory_free
-from vllm_ascend.utils import vllm_version_is
-
-pytestmark = pytest.mark.skipif(
-    not vllm_version_is("0.23.0"),
-    reason="broken on main, fix me.",
-)
 
 
 @wait_until_npu_memory_free()
