+++ tests/ut/eplb/core/a2/test_eplb_utils.py
@@ -9,7 +9,6 @@ from vllm.model_executor.layers.fused_moe.config import FusedMoEConfig, FusedMoE
 
 from vllm_ascend.ascend_config import init_ascend_config
 from vllm_ascend.eplb.core.eplb_utils import generate_log2phy_map, init_eplb_config
-from vllm_ascend.utils import vllm_version_is
 # isort: on
 
 
@@ -26,36 +25,19 @@ class TestAscendConfig(unittest.TestCase):
         from vllm.model_executor.layers.fused_moe.config import RoutingMethodType
 
         moe_parallel_config = FusedMoEParallelConfig(2, 0, 1, 2, 1, 1, 1, 1, 1, True, "hccl", enable_eplb=True)
-        if vllm_version_is("0.23.0"):
-            moe_config = FusedMoEConfig(
-                num_experts=8,
-                experts_per_token=8,
-                hidden_dim=8192,
-                intermediate_size_per_partition=5,
-                num_local_experts=8,
-                num_logical_experts=8,
-                activation="silu",
-                device="npu",
-                routing_method=RoutingMethodType.Simulated,
-                moe_parallel_config=moe_parallel_config,
-                in_dtype=torch.float16,
-            )
-        else:
-            from vllm.model_executor.layers.fused_moe.activation import MoEActivation
-
-            moe_config = FusedMoEConfig(
-                num_experts=8,
-                experts_per_token=8,
-                hidden_dim=8192,
-                intermediate_size=10,
-                num_local_experts=8,
-                num_logical_experts=8,
-                activation=MoEActivation.SILU,
-                device="npu",
-                routing_method=RoutingMethodType.Simulated,
-                moe_parallel_config=moe_parallel_config,
-                in_dtype=torch.float16,
-            )
+        moe_config = FusedMoEConfig(
+            num_experts=8,
+            experts_per_token=8,
+            hidden_dim=8192,
+            intermediate_size_per_partition=5,
+            num_local_experts=8,
+            num_logical_experts=8,
+            activation="silu",
+            device="npu",
+            routing_method=RoutingMethodType.Simulated,
+            moe_parallel_config=moe_parallel_config,
+            in_dtype=torch.float16,
+        )
         moe_config.supports_eplb = True
         self.vllm_config = vllm_config
         self.moe_config = moe_config
