+++ tests/ut/quantization/test_compressed_tensors_config.py
@@ -1,6 +1,5 @@
 from unittest.mock import MagicMock, patch
 
-import pytest
 from vllm.model_executor.layers.attention import Attention
 from vllm.model_executor.layers.fused_moe import FusedMoE
 from vllm.model_executor.layers.linear import RowParallelLinear, UnquantizedLinearMethod
@@ -11,7 +10,7 @@ from vllm_ascend.ops.fused_moe.fused_moe import AscendUnquantizedFusedMoEMethod
 from vllm_ascend.quantization.compressed_tensors_config import AscendCompressedTensorsConfig
 from vllm_ascend.quantization.method_adapters import AscendFusedMoEMethod, AscendLinearMethod
 from vllm_ascend.quantization.methods import AscendW8A8DynamicFusedMoEMethod, AscendW8A8DynamicLinearMethod
-from vllm_ascend.utils import COMPRESSED_TENSORS_METHOD, vllm_version_is
+from vllm_ascend.utils import COMPRESSED_TENSORS_METHOD
 
 
 class TestAscendCompressedTensorsQuanType(TestBase):
@@ -95,10 +94,6 @@ class TestAscendCompressedTensorsConfigGetQuantMethod(TestBase):
         self.assertEqual(layer.ascend_quant_method, COMPRESSED_TENSORS_METHOD)
         self.assertTrue(isinstance(result, UnquantizedLinearMethod))
 
-    @pytest.mark.skipif(
-        not vllm_version_is("0.23.0"),
-        reason="Legacy FusedMoE quant method UT is only for vLLM 0.23.0.",
-    )
     @patch("vllm_ascend.quantization.methods.AscendW8A8DynamicFusedMoEMethod.__init__")
     def test_get_moe_quant_method(self, mock_method):
         mock_method.return_value = None
@@ -109,10 +104,6 @@ class TestAscendCompressedTensorsConfigGetQuantMethod(TestBase):
         self.assertTrue(isinstance(result, AscendFusedMoEMethod))
         self.assertTrue(isinstance(layer.scheme, AscendW8A8DynamicFusedMoEMethod))
 
-    @pytest.mark.skipif(
-        not vllm_version_is("0.23.0"),
-        reason="Legacy FusedMoE quant method UT is only for vLLM 0.23.0.",
-    )
     @patch("vllm_ascend.ops.fused_moe.fused_moe.AscendUnquantizedFusedMoEMethod.__init__")
     @patch("vllm_ascend.quantization.compressed_tensors_config.should_ignore_layer")
     def test_get_moe_unquantized_method(self, mock_ignore_layer, mock_method):
