+++ vllm_ascend/ops/fused_moe/moe_runtime_args.py
@@ -77,7 +77,6 @@ from vllm_ascend.ops.fused_moe.moe_stage_params import (
     MoERoutingParams,
 )
 from vllm_ascend.quantization.quant_type import QuantType
-from vllm_ascend.utils import vllm_version_is
 
 
 def _build_mxfp_params(
@@ -146,12 +145,8 @@ def build_fused_experts_input(
     w2_scale_bias: list[torch.Tensor] | torch.Tensor | None = None,
     w1_offset: torch.Tensor | None = None,
     w2_offset: torch.Tensor | None = None,
-    swiglu_limit: float | None = 0.0,
+    swiglu_limit: float = 0.0,
 ) -> MoEFusedExpertsInput:
-    if not vllm_version_is("0.23.0") and swiglu_limit is None:
-        swiglu_limit = 0.0
-    assert swiglu_limit is not None
-
     return MoEFusedExpertsInput(
         hidden_states=hidden_states,
         topk_weights=topk_weights,
